feat: new agnostic reflection engine.
This commit adds a new reflection engine that is agnostic, allowing us to be able to use it for other future classes such as user and message context menus, etc. that are not just `NexusCommand`.

The new reflection engine is also written in Kotlin, but uses the same Java Reflection, enabling more readability and cleanliness.

This also adds a new `@Uuid` annotation that is used to figure out which field is the origin Uuid field of the reference class.